pyproject.toml
[project]
name = "adk-prac"
version = "0.1.0"
description = "Google ADK practice with Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/adk_prac"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
